requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alabaster==0.7.13
arrow==1.3.0
astroid==3.0.2
asttokens==2.4.1
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==23.1.0
autopep8==2.0.4
Babel==2.14.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
black==23.12.0
bleach==6.1.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.0
cookiecutter==2.5.0
cryptography==41.0.7
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.7
docstring-to-markdown==0.13
docutils==0.20.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.68.0
fastjsonschema==2.19.0
filelock==3.13.1
flake8==6.1.0
frozenlist==1.4.1
fsspec==2023.12.2
huggingface-hub==0.19.4
idna==3.6
imagesize==1.4.1
importlib-metadata==7.0.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.27.1
ipython==8.18.1
isort==5.13.2
jaraco.classes==3.3.0
jedi==0.19.1
jeepney==0.8.0
jellyfish==1.0.3
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyterlab_pygments==0.3.0
keyring==24.3.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.4.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.13.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.21.1
numpydoc==1.6.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==4.1.0
pluggy==1.3.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
pydantic==1.10.13
pydocstyle==6.3.0
pyflakes==3.1.0
Pygments==2.17.2
pylint==3.0.3
pylint-venv==3.0.3
pyls-spyder==0.4.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
python-dateutil==2.8.2
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.9.0
python-slugify==8.0.1
pytoolconfig==1.2.6
pytz==2023.3.post1
pyxdg==0.28
PyYAML==6.0.1
pyzmq==25.1.2
QDarkStyle==3.2.3
qstylizer==0.2.2
QtAwesome==1.3.0
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
rich==13.7.0
rope==1.11.0
rpds-py==0.15.2
Rtree==1.1.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.7.0
SecretStorage==3.3.3
six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
spyder==5.5.0
spyder-kernels==2.5.0
stack-data==0.6.3
starlette==0.14.2
sympy==1.12
tabulate==0.9.0
text-unidecode==1.3
textdistance==4.6.0
threadpoolctl==3.2.0
three-merge==0.1.1
tinycss2==1.2.1
tokenizers==0.15.0
tomli==2.0.1
tomlkit==0.12.3
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.62.3
traitlets==5.14.0
transformers==4.36.2
triton==2.1.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.1.0
watchdog==3.0.0
wcwidth==0.2.12
webencodings==0.5.1
whatthepatch==1.0.5
wurlitzer==3.0.3
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
